Rename: components 폴더 생성 및 폴더 구조 변경
