+++ components/src/lib/components/modalMessage/ModalMessage.tsx
@@ -1,5 +1,5 @@
 import "./ModalMessage.css";
-import { ModalMessageProps } from "../type/modal.type";
+import { ModalMessageProps } from "../../type/modal.type";
 import { PropsWithChildren } from "react";
 
 const ModalMessage = ({ children, style = {}, className = "", ...rest }: PropsWithChildren<ModalMessageProps>) => {
